Updated shield + revert material
My OS is borked (see below), so I have to put this on hold for a bit. I've reverted the WIP _material integration with 3d_armor set bonus detection, and changed the shield overlay colorization from orange to black to fit the mod author's color scheme. The armor is back to looking and acting normal, but the shield will throw a non-critical error when you try to equip it because there are no model or textures yet. All that will happen is your character will turn pink, but you will still get all the effects and protection.

I will be merging this init and the readme to my fork's master if i can figure that out, hopefully without also sending it upstream just yet.